+++ R/util.R
@@ -106,18 +106,19 @@ integrate.xy <- function(x,fx, a,b, use.spline = TRUE, xtol = 2e-8)
     r/2
 }
 
-buildShinyElements <- function(questions, itemnames){
+buildShinyElements <- function(questions, itemnames, customTypes){
     J <- length(questions$Type)
-    if(!all(sapply(questions[names(questions) != 'Question'], is.character))) 
+    if(!all(sapply(questions[names(questions) != 'Rendered_Question'], is.character))) 
         stop('Only character classes are supported in questions input', call.=FALSE)
     if(is.null(itemnames)) itemnames <- paste0('Item.', 1L:J)
     names <- names(questions)
-    Qs_char <- questions$Question
+    Qs_char <- questions$Rendered_Question
     Type <- questions$Type
     if(is.null(Qs_char) && !any(questions$Type == 'slider')) 
         stop('Question column not specified', call.=FALSE)
     if(is.null(Type)) stop('Type column not specified', call.=FALSE)
-    if(!all(Type %in% c('radio', 'select', 'text', 'slider', 'checkbox', 'none')))
+    if(!all(Type %in% c('radio', 'select', 'text', 'textArea', 'slider', 'checkbox', 'none',
+                        names(customTypes))))
         stop('Type input in shiny_questions contains invalid arguments', call.=FALSE)
     Qs <- vector('list', J)
     choices <- data.frame(questions[grepl('Option', names)], stringsAsFactors = FALSE)
@@ -144,6 +145,16 @@ buildShinyElements <- function(questions, itemnames){
             placeholder <- questions$placeholder[i]
             Qs[[i]] <- textInput(inputId = itemnames[i], label='', value = '',
                                  width=width, placeholder=placeholder)
+        } else if(Type[i] == 'textArea'){
+            width <- questions$width[i]
+            height <- questions$height[i]
+            cols <- questions$cols[i] 
+            rows <- questions$rows[i]
+            resize <- questions$resize[i]
+            placeholder <- questions$placeholder[i]
+            Qs[[i]] <- textAreaInput(inputId = itemnames[i], label='', value = '',
+                                 width=width, height=height, placeholder=placeholder,
+                                 cols=cols, rows=rows, resize=resize)
         } else if(Type[i] == 'slider'){
             VALUE <- as.numeric(ifelse(is.null(questions$value[i]), questions$min[i], questions$value[i]))
             MIN <- as.numeric(questions$min[i])
@@ -165,6 +176,12 @@ buildShinyElements <- function(questions, itemnames){
             inline <- if(is.null(questions$inline[i])) FALSE else as.logical(questions$inline[i])
             Qs[[i]] <- checkboxGroupInput(inputId = itemnames[i], label='', choices = cs,
                                           width=width, inline=inline)
+        } else if(Type[i] %in% names(customTypes)){
+            nm <- names(customTypes)[names(customTypes) == Type[i]]
+            df_row <- as.data.frame(lapply(questions, function(x, ind) x[[ind]], ind=i), 
+                                    stringsAsFactors = FALSE)
+            df_row$Rendered_Question <- NULL
+            Qs[[i]] <- customTypes[[nm]](inputId = itemnames[i], df_row=df_row)
         } else if(Type[i] == 'none'){
             next
         }
@@ -172,7 +189,11 @@ buildShinyElements <- function(questions, itemnames){
     pick <- as.data.frame(questions[grepl('Answer', names),drop=FALSE], stringsAsFactors = FALSE)
     if(length(pick)){
         item_answers <- split(pick, 1:nrow(pick))
-        item_answers <- lapply(item_answers, na.omit)
+        item_answers <- lapply(item_answers, function(x){
+            out <- na.omit(as.character(x))
+            if(!length(out)) out <- NA
+            out
+        })
     } else item_answers <- NULL
     if(length(Qs) != J) stop('Questions have not been properly defined!', call.=FALSE)
     ret <- list(questions=Qs, item_answers=item_answers, item_options=choices_list)